Add API response structure, base controller, and booking mapper; implement booking service methods

+++ booking-service/src/main/java/com/lamnd/dto/response/BookingResponse.java
@@ -15,6 +15,6 @@ public record BookingResponse(
         LocalDateTime startTime,
         LocalDateTime endTime,
         BookingStatus status,
-        int totalPrice
+        Double totalPrice
 ) {
 }
